requirements.txt
Python-docx
pandas
openpyxl
xlrd
PyMuPDF
tiktoken
chromadb
sentence-transformers
argostranslate
transformers
rouge-score